requirements.txt
pyaudio
groq
elevenlabs
langchain
langchain-groq
langchain-community
numpy
pydub
python-dotenv
flask
certifi
paho-mqtt
flask-cors